Partial WebSocket implementation

+++ app/src/main/java/com/github/poundr/network/model/AuthResponse.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class AuthResponse(
-    @Json(name = "profileId") val profileId: Int?,
+    @Json(name = "profileId") val profileId: Long?,
     @Json(name = "xmppToken") val xmppToken: String?,
     @Json(name = "sessionId") val sessionId: String?,
     @Json(name = "authToken") val authToken: String?,
